Remove skeletons from sidebar for now
